Share generic iterator tests with PushQueue.
