+++ test/unit/iterators.test.js
@@ -4,7 +4,8 @@ import { iteratorFinally, CancelableGenerator, pipeline } from '../../src/utils/
 import { Defer } from '../../src/utils'
 import PushQueue from '../../src/utils/PushQueue'
 
-const expected = [1, 2, 3, 4, 5, 6, 7, 8]
+import IteratorTest, { expected, MAX_ITEMS } from './IteratorTest'
+
 const WAIT = 20
 
 async function* generate(items = expected) {
@@ -15,189 +16,6 @@ async function* generate(items = expected) {
         await wait(WAIT * 0.1)
     }
     await wait(WAIT * 0.1)
-}
-
-const MAX_ITEMS = 3
-
-function IteratorTest(name, fn) {
-    describe(`${name} IteratorTest`, () => {
-        it('runs to completion', async () => {
-            const received = []
-            const itr = fn()
-            for await (const msg of itr) {
-                received.push(msg)
-            }
-            expect(received).toEqual(expected)
-        })
-
-        it('can return in finally', async () => {
-            const received = []
-            const itr = (async function* Outer() {
-                const innerItr = fn()[Symbol.asyncIterator]()
-                try {
-                    yield* innerItr
-                } finally {
-                    await innerItr.return() // note itr.return would block
-                }
-            }())
-
-            for await (const msg of itr) {
-                received.push(msg)
-                if (received.length === MAX_ITEMS) {
-                    break
-                }
-            }
-            expect(received).toEqual(expected.slice(0, MAX_ITEMS))
-        })
-
-        it('can return mid-iteration', async () => {
-            const received = []
-            for await (const msg of fn()) {
-                received.push(msg)
-                if (received.length === MAX_ITEMS) {
-                    break
-                }
-            }
-            expect(received).toEqual(expected.slice(0, MAX_ITEMS))
-        })
-
-        it('can throw mid-iteration', async () => {
-            const received = []
-            const err = new Error('expected err')
-            await expect(async () => {
-                for await (const msg of fn()) {
-                    received.push(msg)
-                    if (received.length === MAX_ITEMS) {
-                        throw err
-                    }
-                }
-            }).rejects.toThrow(err)
-            expect(received).toEqual(expected.slice(0, MAX_ITEMS))
-        })
-
-        it('can .throw() mid-iteration', async () => {
-            const received = []
-            const err = new Error('expected err')
-            await expect(async () => {
-                const it = fn()
-                for await (const msg of it) {
-                    received.push(msg)
-                    if (received.length === MAX_ITEMS) {
-                        await it.throw(err)
-                    }
-                }
-            }).rejects.toThrow(err)
-            expect(received).toEqual(expected.slice(0, MAX_ITEMS))
-        })
-
-        it('can .throw() mid-iteration and run delayed finally block', async () => {
-            const received = []
-            const err = new Error('expected err')
-            const onFinally = jest.fn()
-            const onFinallyDelayed = jest.fn(async () => {
-                await wait(100)
-                return onFinally()
-            })
-
-            await expect(async () => {
-                const it = fn()
-                try {
-                    for await (const msg of it) {
-                        received.push(msg)
-                        if (received.length === MAX_ITEMS) {
-                            await it.throw(err)
-                        }
-                    }
-                } finally {
-                    await onFinallyDelayed()
-                }
-            }).rejects.toThrow(err)
-            expect(received).toEqual(expected.slice(0, MAX_ITEMS))
-            expect(onFinallyDelayed).toHaveBeenCalledTimes(1)
-            expect(onFinally).toHaveBeenCalledTimes(1)
-        })
-
-        it('can throw in delayed finally block', async () => {
-            const received = []
-            const err = new Error('expected err')
-            const onFinally = jest.fn()
-            const onFinallyDelayed = jest.fn(async () => {
-                await wait(50)
-                await onFinally()
-                throw err
-            })
-
-            await expect(async () => {
-                const it = fn()
-                try {
-                    for await (const msg of it) {
-                        received.push(msg)
-                        if (received.length === MAX_ITEMS) {
-                            break
-                        }
-                    }
-                } finally {
-                    await onFinallyDelayed()
-                }
-            }).rejects.toThrow(err)
-            expect(received).toEqual(expected.slice(0, MAX_ITEMS))
-            expect(onFinallyDelayed).toHaveBeenCalledTimes(1)
-            expect(onFinally).toHaveBeenCalledTimes(1)
-        })
-
-        it('throws parent mid-iteration', async () => {
-            const received = []
-            const err = new Error('expected err')
-            async function* parentGen() {
-                for await (const msg of fn()) {
-                    yield msg
-                    if (received.length === MAX_ITEMS) {
-                        throw err
-                    }
-                }
-            }
-            await expect(async () => {
-                for await (const msg of parentGen()) {
-                    received.push(msg)
-                }
-            }).rejects.toThrow(err)
-            expect(received).toEqual(expected.slice(0, MAX_ITEMS))
-        })
-
-        it('can throw before iterating', async () => {
-            const received = []
-            const itr = fn()[Symbol.asyncIterator]()
-            const err = new Error('expected err')
-
-            await expect(async () => {
-                await itr.throw(err)
-            }).rejects.toThrow(err)
-
-            // does not throw
-            for await (const msg of itr) {
-                received.push(msg)
-            }
-            expect(received).toEqual([])
-        })
-
-        it('can return before iterating', async () => {
-            const itr = fn()[Symbol.asyncIterator]()
-            await itr.return()
-            const received = []
-            for await (const msg of itr) {
-                received.push(msg)
-            }
-            expect(received).toEqual([])
-        })
-
-        it('can queue next calls', async () => {
-            const itr = fn()[Symbol.asyncIterator]()
-            const tasks = expected.map(async () => itr.next())
-            const received = await Promise.all(tasks)
-            expect(received.map(({ value }) => value)).toEqual(expected)
-            await itr.return()
-        })
-    })
 }
 
 describe('Iterator Utils', () => {
